privy: refactor type filenames
