+++ code/ext/ext.lib.privy/src/common/index.ts
@@ -1,6 +1,5 @@
-import type * as t from './types';
+export type * as t from './t';
 
 export * from '../index.pkg.mjs';
 export * from './env';
 export * from './libs';
-export type { t };
