+++ code/ext/ext.lib.privy/src/common/t.ts
