+++ code/ext/ext.lib.privy/src/ui/common/index.ts
@@ -1,8 +1,7 @@
-import type * as t from './types';
+export type * as t from './t';
 
 export * from '../../common';
-export * from './libs.mjs';
-export { type t };
+export * from './libs.js';
 
 export { Balance, Chain, Wallet } from '../../evm';
 export { Icons } from '../Icons';
