+++ code/ext/ext.lib.privy/src/ui/common/t.ts
@@ -13,4 +13,4 @@ export type {
 /**
  * @local
  */
-export type * from '../../common/types';
+export type * from '../../common/t';
